Cleanup REST Api types and harden property naming-convention
