+++ packages/demo-protobuf/src/registry.spec.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/naming-convention */
 import { assert } from "@cosmjs/utils";
 
 import { MsgDemo as MsgDemoType } from "./demo";
