+++ packages/demo-protobuf/types/decorator.d.ts
@@ -1,12 +1,12 @@
-import { Constructor, Message, TypeDecorator } from "protobufjs";
+import { Constructor, FieldDecorator, Message, TypeDecorator } from "protobufjs";
 import { Registry } from "./registry";
 export declare function CosmosMessage(registry: Registry, typeUrl: string): TypeDecorator<any>;
 export declare const CosmosField: {
-  Boolean: (id: number) => import("protobufjs").FieldDecorator;
-  String: (id: number) => import("protobufjs").FieldDecorator;
-  Bytes: (id: number) => import("protobufjs").FieldDecorator;
-  Int64: (id: number) => import("protobufjs").FieldDecorator;
-  Uint64: (id: number) => import("protobufjs").FieldDecorator;
-  RepeatedString: (id: number) => import("protobufjs").FieldDecorator;
-  Nested: (id: number, ctor: Constructor<Message<{}>>) => import("protobufjs").FieldDecorator;
+  Boolean: (id: number) => FieldDecorator;
+  String: (id: number) => FieldDecorator;
+  Bytes: (id: number) => FieldDecorator;
+  Int64: (id: number) => FieldDecorator;
+  Uint64: (id: number) => FieldDecorator;
+  RepeatedString: (id: number) => FieldDecorator;
+  Nested: (id: number, ctor: Constructor<Message>) => FieldDecorator;
 };
