+++ packages/sdk38/src/cosmosclient.searchtx.spec.ts
@@ -43,7 +43,9 @@ describe("CosmosClient.searchTx", () => {
         const sendMsg: MsgSend = {
           type: "cosmos-sdk/MsgSend",
           value: {
+            // eslint-disable-next-line @typescript-eslint/naming-convention
             from_address: faucet.address,
+            // eslint-disable-next-line @typescript-eslint/naming-convention
             to_address: recipient,
             amount: transferAmount,
           },
