+++ packages/sdk38/src/cosmosclient.spec.ts
@@ -199,7 +199,9 @@ describe("CosmosClient", () => {
       const sendMsg: MsgSend = {
         type: "cosmos-sdk/MsgSend",
         value: {
+          // eslint-disable-next-line @typescript-eslint/naming-convention
           from_address: faucet.address,
+          // eslint-disable-next-line @typescript-eslint/naming-convention
           to_address: makeRandomAddress(),
           amount: [
             {
