+++ packages/sdk38/src/msgs.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/naming-convention */
 import { Coin } from "./coins";
 
 export interface Msg {
