+++ packages/sdk38/src/signature.ts
@@ -17,6 +17,7 @@ export function encodeSecp256k1Signature(pubkey: Uint8Array, signature: Uint8Arr
   }
 
   return {
+    // eslint-disable-next-line @typescript-eslint/naming-convention
     pub_key: encodeSecp256k1Pubkey(pubkey),
     signature: toBase64(signature),
   };
