+++ packages/sdk38/src/signingcosmosclient.ts
@@ -84,7 +84,9 @@ export class SigningCosmosClient extends CosmosClient {
     const sendMsg: MsgSend = {
       type: "cosmos-sdk/MsgSend",
       value: {
+        // eslint-disable-next-line @typescript-eslint/naming-convention
         from_address: this.senderAddress,
+        // eslint-disable-next-line @typescript-eslint/naming-convention
         to_address: recipientAddress,
         amount: transferAmount,
       },
